# api/openapispec/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/workspaces/configs/validate": {
            "post": {
                "description": "Validate the configurations in the specified workspace",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "Validate workspace configurations",
                "operationId": "validateWorkspaceConfigs",
                "parameters": [
                    {
                        "description": "Workspace configurations to be validated",
                        "name": "workspace",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.WorkspaceConfigs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/request.WorkspaceConfigs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
        "/api/v1/workspaces/{id}/configs": {
            "get": {
                "description": "Get configurations in the specified workspace",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "get workspace configurations",
                "operationId": "getWorkspaceConfigs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workspace ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/request.WorkspaceConfigs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "put": {
                "description": "Update the configurations in the specified workspace",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "Update workspace configurations",
                "operationId": "updateWorkspaceConfigs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workspace ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated workspace configurations",
                        "name": "workspace",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.WorkspaceConfigs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/request.WorkspaceConfigs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/v1/workspaces/{id}/configs/mod-deps": {
            "post": {
                "description": "Create the module dependencies in kcl.mod of the specified workspace",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "Create the module dependencies of the workspace",
                "operationId": "createWorkspaceModDeps",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workspace ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.AWSProvider": {
            "type": "object",
            "properties": {
                "profile": {
                    "description": "The profile to be used to interact with AWS Secrets Manager.\nIf not set, the default profile created with `aws configure` will be used.",
                    "type": "string"
                },
                "region": {
                    "description": "AWS Region to be used to interact with AWS Secrets Manager.\nExamples are us-east-1, us-west-2, etc.",
                    "type": "string"
                }
            }
        },
        "kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.AlicloudProvider": {
            "type": "object",
            "properties": {
                "region": {
                    "description": "Alicloud Region to be used to interact with Alicloud Secrets Manager.\nExamples are cn-beijing, cn-shanghai, etc.",
                    "type": "string"
                }
            }
        },
        "kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.AzureEnvironmentType": {
            "type": "string",
            "enum": [
                "PublicCloud",
                "USGovernmentCloud",
                "ChinaCloud",
                "GermanCloud"
            ],
            "x-enum-varnames": [
                "AzureEnvironmentPublicCloud",
                "AzureEnvironmentUSGovernmentCloud",
                "AzureEnvironmentChinaCloud",
                "AzureEnvironmentGermanCloud"
            ]
        },
        "kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.AzureKVProvider": {
            "type": "object",
            "properties": {
                "environmentType": {
                    "description": "EnvironmentType specifies the Azure cloud environment endpoints to use for connecting and authenticating with Azure.\nBy-default it points to the public cloud AAD endpoint, and the following endpoints are available:\nPublicCloud, USGovernmentCloud, ChinaCloud, GermanCloud\nRef: https://github.com/Azure/go-autorest/blob/main/autorest/azure/environments.go#L152",
                    "allOf": [
                        {
                            "$ref": "#/definitions/kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.AzureEnvironmentType"
                        }
                    ]
                },
                "tenantId": {
                    "description": "TenantID configures the Azure Tenant to send requests to.",
                    "type": "string"
                },
                "vaultUrl": {
                    "description": "Vault Url from which the secrets to be fetched from.",
                    "type": "string"
                }
            }
        },
...
        "kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.Configs": {
            "type": "object",
            "properties": {
                "default": {
                    "description": "Default is default block of the module config.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.GenericConfig"
                        }
                    ]
                }
            }
        },
        "kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.FakeProvider": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.FakeProviderData"
                    }
                }
            }
        },
        "kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.FakeProviderData": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                },
                "valueMap": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.GenericConfig": {
            "type": "object",
            "additionalProperties": {}
        },
        "kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.ModuleConfig": {
            "type": "object",
            "properties": {
                "configs": {
                    "description": "Configs contains all levels of module configs",
                    "allOf": [
                        {
                            "$ref": "#/definitions/kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.Configs"
                        }
                    ]
                },
                "path": {
                    "description": "Path is the path of the module. It can be a local path or a remote URL",
                    "type": "string"
                },
                "version": {
                    "description": "Version is the version of the module.",
                    "type": "string"
                }
            }
        },
        "kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.ModuleConfigs": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.ModuleConfig"
            }
        },
...
        "kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.OnPremisesProvider": {
            "type": "object",
            "properties": {
                "attributes": {
                    "description": "attributes of the provider",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "platform name of the provider",
                    "type": "string"
                }
            }
        },
        "kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.ProviderSpec": {
            "type": "object",
            "properties": {
                "alicloud": {
                    "description": "Alicloud configures a store to retrieve secrets from Alicloud Secrets Manager.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.AlicloudProvider"
                        }
                    ]
                },
                "aws": {
                    "description": "AWS configures a store to retrieve secrets from AWS Secrets Manager.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.AWSProvider"
                        }
                    ]
                },
                "azure": {
                    "description": "Azure configures a store to retrieve secrets from Azure KeyVault.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.AzureKVProvider"
                        }
                    ]
                },
                "fake": {
                    "description": "Fake configures a store with static key/value pairs",
                    "allOf": [
                        {
                            "$ref": "#/definitions/kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.FakeProvider"
                        }
                    ]
                },
                "onpremises": {
                    "description": "Onprem configures a store in on-premises environments",
                    "allOf": [
                        {
                            "$ref": "#/definitions/kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.OnPremisesProvider"
                        }
                    ]
                },
                "vault": {
                    "description": "Vault configures a store to retrieve secrets from HashiCorp Vault.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.VaultProvider"
                        }
                    ]
                },
                "viettelcloud": {
                    "description": "ViettelCloud configures a store to retrieve secrets from ViettelCloud Secrets Manager.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.ViettelCloudProvider"
                        }
                    ]
                }
            }
        },
...
        "kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.SecretStore": {
            "type": "object",
            "properties": {
                "provider": {
                    "$ref": "#/definitions/kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.ProviderSpec"
                }
            }
        },
...
        "kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.VaultKVStoreVersion": {
            "type": "string",
            "enum": [
                "v1",
                "v2"
            ],
            "x-enum-varnames": [
                "VaultKVStoreV1",
                "VaultKVStoreV2"
            ]
        },
        "kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.VaultProvider": {
            "type": "object",
            "properties": {
                "path": {
                    "description": "Path is the mount path of the Vault KV backend endpoint, e.g: \"secret\".",
                    "type": "string"
                },
                "server": {
                    "description": "Server is the target Vault server address to connect, e.g: \"https://vault.example.com:8200\".",
                    "type": "string"
                },
                "version": {
                    "description": "Version is the Vault KV secret engine version. Version can be either \"v1\" or\n\"v2\", defaults to \"v2\".",
                    "allOf": [
                        {
                            "$ref": "#/definitions/kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.VaultKVStoreVersion"
                        }
                    ]
                }
            }
        },
        "kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.ViettelCloudProvider": {
            "type": "object",
            "properties": {
                "cmpURL": {
                    "description": "ViettelCloud CMP URL to be used to interact with ViettelCloud Secrets Manager.\nExamples are https://console.viettelcloud.vn/api/",
                    "type": "string"
                },
                "projectID": {
                    "description": "ProjectID to be used to interact with ViettelCloud Secrets Manager.",
                    "type": "string"
                }
            }
        },
...
        "request.WorkspaceConfigs": {
            "type": "object",
            "properties": {
                "context": {
                    "description": "Context contains workspace-level configurations, such as runtimes, topologies, and metadata, etc.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.GenericConfig"
                        }
                    ]
                },
                "modules": {
                    "description": "Modules are the configs of a set of modules.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.ModuleConfigs"
                        }
                    ]
                },
                "secretStore": {
                    "description": "SecretStore represents a secure external location for storing secrets.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/kusionstack_io_kusion_pkg_apis_api_kusion_io_v1.SecretStore"
                        }
                    ]
                }
            }
        },
...
    }
}